Additional nodiscards and constructor changes
