+++ .vscode/c_cpp_properties.json
@@ -4,13 +4,19 @@
             "name": "Linux",
             "includePath": [
                 "${workspaceFolder}/src",
-                "${workspaceFolder}/pros",
-                "/opt/ros/noetic/include/**"
+                "/opt/ros/noetic/include/**",
+                "${workspaceFolder}/pros"
             ],
             "defines": [],
             "compilerPath": "/usr/bin/gcc",
             "cppStandard": "c++20",
-            "intelliSenseMode": "linux-gcc-x64"
+            "intelliSenseMode": "linux-gcc-x64",
+            "browse": {
+                "path": [
+                    "${workspaceFolder}/src"
+                ],
+                "limitSymbolsToIncludedHeaders": true
+            }
         }
     ],
     "version": 4
