+++ src/hardware/pros_motor.h
@@ -1,26 +1,11 @@
 #pragma once
 #include "constant/constants.h"
+#include "hardware/pros_motor_cartridge.h"
 #include "interface/encoder.h"
 #include "interface/motor.h"
 #include "pros.h"
 
 namespace hardware {
-
-enum class ProsMotorCartridge {
-  // equal to pros value to enable casting
-  kBlueCartridge = pros::E_MOTOR_GEAR_BLUE,   // 600 rpm
-  kGreenCartridge = pros::E_MOTOR_GEAR_GREEN  // 200 rpm
-};
-
-[[nodiscard]] pros::motor_gearset_e_t ConvertProsMotorCartridge(
-    ProsMotorCartridge);
-[[nodiscard]] ProsMotorCartridge ConvertProsGearset(pros::motor_gearset_e_t);
-
-/**
- * A helper function for getting the max rpm of a motor.
- */
-[[nodiscard]] int GetMaxRpm(const ProsMotorCartridge);
-
 /**
  * A class which wraps a single pros::Motor object.
  */
@@ -31,14 +16,21 @@ class ProsMotor : public interface::Motor, public interface::Encoder {
             port_number * (reverse ? -1 : 1),
             ConvertProsMotorCartridge(cartridge)) {}
 
+  ProsMotor(const ProsMotor&) = delete;
+  ProsMotor& operator=(const ProsMotor&) = delete;
+
+  ProsMotor(ProsMotor&&) = default;
+  ProsMotor& operator=(ProsMotor&&) = default;
+  ~ProsMotor() = default;
+
   void Move(int) override;
   void MoveVoltage(int) override;
   void MoveVelocity(float) override;
   void MoveAbsolute(double position, int max_velocity) override;
 
   void ResetEncoder() override;
-  double position() const override;
-  float velocity() const override;
+  [[nodiscard]] double position() const override;
+  [[nodiscard]] float velocity() const override;
 
  private:
   [[nodiscard]] ProsMotorCartridge cartridge() const;
