+++ src/hardware/pros_motor_cartridge.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "pros.h"
+
+namespace hardware {
+enum class ProsMotorCartridge {
+  // equal to pros value to enable casting
+  kBlueCartridge = pros::E_MOTOR_GEAR_BLUE,   // 600 rpm
+  kGreenCartridge = pros::E_MOTOR_GEAR_GREEN  // 200 rpm
+};
+
+[[nodiscard]] pros::motor_gearset_e_t ConvertProsMotorCartridge(
+    ProsMotorCartridge);
+[[nodiscard]] ProsMotorCartridge ConvertProsGearset(pros::motor_gearset_e_t);
+
+/**
+ * A helper function for getting the max rpm of a motor.
+ */
+[[nodiscard]] int GetMaxRpm(const ProsMotorCartridge);
+}  // namespace hardware
