+++ src/interface/controller.h
@@ -4,8 +4,8 @@ enum class ControllerJoystick { kLeftX, kLeftY, kRightX, kRightY };
 
 class Controller {
  public:
-  virtual int analog(ControllerJoystick joystick) const = 0;
-  virtual float voltage(ControllerJoystick joystick) const = 0;
+  [[nodiscard]] virtual int analog(ControllerJoystick joystick) const = 0;
+  [[nodiscard]] virtual float voltage(ControllerJoystick joystick) const = 0;
 
   virtual ~Controller() = default;
 };
